Fixing database entries and adding historical claim view
