+++ frontend/app/database/old_page.tsx
@@ -1,0 +1,253 @@
+'use client'
+import { Fragment, useState } from 'react'
+import { Dialog, Disclosure, Menu, Transition } from '@headlessui/react'
+import { XMarkIcon } from '@heroicons/react/24/outline'
+import { ChevronDownIcon, FunnelIcon, MinusIcon, PlusIcon, Squares2X2Icon } from '@heroicons/react/20/solid'
+import Link from "next/link";
+
+const sortOptions = [
+    { name: 'Most Popular', href: '#', current: true },
+    { name: 'Best Rating', href: '#', current: false },
+    { name: 'Newest', href: '#', current: false },
+    { name: 'Price: Low to High', href: '#', current: false },
+    { name: 'Price: High to Low', href: '#', current: false },
+]
+// const subCategories = [
+//   { name: 'Totes', href: '#' },
+// ]
+const filters = [
+    {
+        id: 'amount',
+        name: 'Amount Applied For',
+        options: [
+            { value: '0-10K', label: '0-10K', checked: false },
+            { value: '10K-100K', label: '10K-100K', checked: false },
+            { value: '100K+', label: '100K+', checked: false },
+        ],
+    },
+    {
+        id: 'time',
+        name: 'Time',
+        options: [
+            { value: 'This month', label: 'This month', checked: false },
+            { value: 'Last 6 months', label: 'Last 6 months', checked: false },
+            { value: 'This year', label: 'This year', checked: false },
+            { value: 'More than a year', label: 'More than a year', checked: false },
+        ],
+    }
+]
+
+function classNames(...classes: any[]) {
+    return classes.filter(Boolean).join(' ')
+}
+
+export default function Database() {
+    const [mobileFiltersOpen, setMobileFiltersOpen] = useState(false)
+
+    return (
+        <main className="flex h-screen w-full flex-col items-center justify-between">
+            <div className="flex w-full p-4 items-center border-b border-white">
+                <div className="flex w-full items-center justify-evenly p-2 gap-8">
+                    <Link
+                        className="py-2 border rounded w-full px-4 text-center"
+                        href={"/inbox"}>
+                        Inbox
+                    </Link>
+                    <div
+                        className="py-2 border rounded w-full px-4 bg-green-50 text-black text-center font-bold">
+                        Database
+                    </div>
+                </div>
+                <div className="flex">
+                    <div className="rounded-full bg-green-300 w-10 h-10" />
+                </div>
+            </div>
+            <div className="w-full h-full flex">
+                <div className="w-full h-full text-white p-4">
+                    <h1 className="text-3xl font-bold">Historical Claims</h1>
+                    <div className="bg-white">
+                        <div>
+                            {/* Mobile filter dialog */}
+                            <Transition.Root show={mobileFiltersOpen} as={Fragment}>
+                                <Dialog as="div" className="relative z-40 lg:hidden" onClose={setMobileFiltersOpen}>
+                                    <Transition.Child
+                                        as={Fragment}
+                                        enter="transition-opacity ease-linear duration-300"
+                                        enterFrom="opacity-0"
+                                        enterTo="opacity-100"
+                                        leave="transition-opacity ease-linear duration-300"
+                                        leaveFrom="opacity-100"
+                                        leaveTo="opacity-0"
+                                    >
+                                        <div className="fixed inset-0 bg-black bg-opacity-25" />
+                                    </Transition.Child>
+
+                                    <div className="fixed inset-0 z-40 flex">
+                                        <Transition.Child
+                                            as={Fragment}
+                                            enter="transition ease-in-out duration-300 transform"
+                                            enterFrom="translate-x-full"
+                                            enterTo="translate-x-0"
+                                            leave="transition ease-in-out duration-300 transform"
+                                            leaveFrom="translate-x-0"
+                                            leaveTo="translate-x-full"
+                                        >
+                                            <Dialog.Panel className="relative ml-auto flex h-full w-full max-w-xs flex-col overflow-y-auto bg-white py-4 pb-12 shadow-xl">
+                                                <div className="flex items-center justify-between px-4">
+                                                    <h2 className="text-lg font-medium text-gray-900">Filters</h2>
+                                                    <button
+                                                        type="button"
+                                                        className="-mr-2 flex h-10 w-10 items-center justify-center rounded-md bg-white p-2 text-gray-400"
+                                                        onClick={() => setMobileFiltersOpen(false)}
+                                                    >
+                                                        <span className="sr-only">Close menu</span>
+                                                        <XMarkIcon className="h-6 w-6" aria-hidden="true" />
+                                                    </button>
+                                                </div>
+
+                                                {/* Filters */}
+                                                <form className="mt-4 border-t border-gray-200">
+                                                    <h3 className="sr-only">Categories</h3>
+                                                    {filters.map((section) => (
+                                                        <Disclosure as="div" key={section.id} className="border-t border-gray-200 px-4 py-6">
+                                                            {({ open }) => (
+                                                                <>
+                                                                    <h3 className="-mx-2 -my-3 flow-root">
+                                                                        <Disclosure.Button className="flex w-full items-center justify-between bg-white px-2 py-3 text-gray-400 hover:text-gray-500">
+                                                                            <span className="font-medium text-gray-900">{section.name}</span>
+                                                                            <span className="ml-6 flex items-center">
+                                            {open ? (
+                                                <MinusIcon className="h-5 w-5" aria-hidden="true" />
+                                            ) : (
+                                                <PlusIcon className="h-5 w-5" aria-hidden="true" />
+                                            )}
+                                            </span>
+                                                                        </Disclosure.Button>
+                                                                    </h3>
+                                                                    <Disclosure.Panel className="pt-6">
+                                                                        <div className="space-y-6">
+                                                                            {section.options.map((option, optionIdx) => (
+                                                                                <div key={option.value} className="flex items-center">
+                                                                                    <input
+                                                                                        id={`filter-mobile-${section.id}-${optionIdx}`}
+                                                                                        name={`${section.id}[]`}
+                                                                                        defaultValue={option.value}
+                                                                                        type="checkbox"
+                                                                                        defaultChecked={option.checked}
+                                                                                        className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                                                                                    />
+                                                                                    <label
+                                                                                        htmlFor={`filter-mobile-${section.id}-${optionIdx}`}
+                                                                                        className="ml-3 min-w-0 flex-1 text-gray-500"
+                                                                                    >
+                                                                                        {option.label}
+                                                                                    </label>
+                                                                                </div>
+                                                                            ))}
+                                                                        </div>
+                                                                    </Disclosure.Panel>
+                                                                </>
+                                                            )}
+                                                        </Disclosure>
+                                                    ))}
+                                                </form>
+                                            </Dialog.Panel>
+                                        </Transition.Child>
+                                    </div>
+                                </Dialog>
+                            </Transition.Root>
+
+                            <main className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
+                                <div className="flex items-baseline justify-between border-b border-gray-200 pb-6">
+                                    <h1 className="text-4xl font-bold tracking-tight text-gray-900">Historical Claims</h1>
+
+                                    <div className="flex items-center">
+                                        <button type="button" className="-m-2 ml-5 p-2 text-gray-400 hover:text-gray-500 sm:ml-7">
+                                            <span className="sr-only">View grid</span>
+                                            <Squares2X2Icon className="h-5 w-5" aria-hidden="true" />
+                                        </button>
+                                        <button
+                                            type="button"
+                                            className="-m-2 ml-4 p-2 text-gray-400 hover:text-gray-500 sm:ml-6 lg:hidden"
+                                            onClick={() => setMobileFiltersOpen(true)}
+                                        >
+                                            <span className="sr-only">Filters</span>
+                                            <FunnelIcon className="h-5 w-5" aria-hidden="true" />
+                                        </button>
+                                    </div>
+                                </div>
+
+                                <section aria-labelledby="products-heading" className="pb-24 pt-6">
+                                    <h2 id="products-heading" className="sr-only">
+                                        Products
+                                    </h2>
+
+                                    <div className="grid grid-cols-1 gap-x-8 gap-y-10 lg:grid-cols-4">
+                                        {/* Filters */}
+                                        <form className="hidden lg:block">
+                                            <h3 className="sr-only">Categories</h3>
+                                            {/* <ul role="list" className="space-y-4  pb-6 text-sm font-medium text-gray-900">
+                            {subCategories.map((category) => (
+                                <li key={category.name}>
+                                <a href={category.href}>{category.name}</a>
+                                </li>
+                            ))}
+                            </ul> */}
+
+                                            {filters.map((section) => (
+                                                <Disclosure as="div" key={section.id} className="border-b border-gray-200 py-6">
+                                                    {({ open }) => (
+                                                        <>
+                                                            <h3 className="-my-3 flow-root">
+                                                                <Disclosure.Button className="flex w-full items-center justify-between bg-white py-3 text-sm text-gray-400 hover:text-gray-500">
+                                                                    <span className="font-medium text-gray-900">{section.name}</span>
+                                                                    <span className="ml-6 flex items-center">
+                                        {open ? (
+                                            <MinusIcon className="h-5 w-5" aria-hidden="true" />
+                                        ) : (
+                                            <PlusIcon className="h-5 w-5" aria-hidden="true" />
+                                        )}
+                                        </span>
+                                                                </Disclosure.Button>
+                                                            </h3>
+                                                            <Disclosure.Panel className="pt-6">
+                                                                <div className="space-y-4">
+                                                                    {section.options.map((option, optionIdx) => (
+                                                                        <div key={option.value} className="flex items-center">
+                                                                            <input
+                                                                                id={`filter-${section.id}-${optionIdx}`}
+                                                                                name={`${section.id}[]`}
+                                                                                defaultValue={option.value}
+                                                                                type="checkbox"
+                                                                                defaultChecked={option.checked}
+                                                                                className="h-4 w-4 rounded border-gray-300 text-indigo-600 focus:ring-indigo-500"
+                                                                            />
+                                                                            <label
+                                                                                htmlFor={`filter-${section.id}-${optionIdx}`}
+                                                                                className="ml-3 text-sm text-gray-600"
+                                                                            >
+                                                                                {option.label}
+                                                                            </label>
+                                                                        </div>
+                                                                    ))}
+                                                                </div>
+                                                            </Disclosure.Panel>
+                                                        </>
+                                                    )}
+                                                </Disclosure>
+                                            ))}
+                                        </form>
+
+                                        {/* Product grid */}
+                                        <div className="lg:col-span-3">{/* Your content */}</div>
+                                    </div>
+                                </section>
+                            </main>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </main>
+
+    )
+}
